Remove wx-specific code from pyface/utils.
