+++ pyface/util/grid/grid_column.py
@@ -1,5 +1,4 @@
-#------------------------------------------------------------------------------
-# Copyright (c) 2005, Enthought, Inc.
+# Copyright (c) 2017, Enthought, Inc.
 # All rights reserved.
 #
 # This software is provided without warranty under the terms of the BSD
@@ -7,27 +6,12 @@
 # under the conditions described in the aforementioned license.  The license
 # is also available online at http://www.enthought.com/licenses/BSD.txt
 # Thanks for using Enthought open source!
-#
-# Author: Enthought, Inc.
-# Description: <Enthought pyface package component>
-#------------------------------------------------------------------------------
-""" A description of a column in a grid. """
 
+import logging
 
-# Enthought library imports.
-from traits.api import Bool, HasTraits, Str
+logger = logging.getLogger(__name__)
+logger.warning(
+    'DEPRECATED: pyface.util.grid.grid_column, use pyface.wx.grid.grid_column instead. '
+    'Will be removed in Pyface 7.')
 
-
-class GridColumn(HasTraits):
-    """ A description of a column in a grid. """
-
-    # Column header.
-    label = Str
-
-    # Type name of data allowed in the column.
-    type = Str
-
-    # Is the column read-only?
-    readonly = Bool(False)
-
-#### EOF ######################################################################
+from pyface.wx.grid.grid_column import *
